+++ ICEngine/src/Graphics/Texture.h
@@ -5,11 +5,11 @@
 
 #include "TypesGraphics.h"
 
-unsigned int ICE_TextureManager_Insert();
-void ICE_TextureManager_Destroy(const unsigned int man);
+size_t ICE_TextureManager_Insert();
+void ICE_TextureManager_Destroy(const size_t man);
 void ICE_TextureManager_DestroyAll();
-unsigned int ICE_Texture_Insert(int manager, char* path);
+size_t ICE_Texture_Insert(int manager, char* path);
 void ICE_Texture_Destroy(ICE_Texture *tex);
-ICE_Texture * ICE_Texture_Get(unsigned int man, unsigned int nb);
+ICE_Texture * ICE_Texture_Get(size_t man, size_t nb);
 
 #endif
